+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## v0.1.2
 
+- new features
+  - convert {:ok, a, b} tuples into {:ok, {a, b}}, same for :error
+
 - bugs fixed
   - restore correct ok? and err? behavior, fix the test
+
+- backwards incompatible
+  - rename rexult to rexult!
+  - rename is_result! to is_rexult!
+  - rename unbreak to unbreak!
 
 ## v0.1.1